resolve stash conflict
