+++ apple/OmnivoreKit/Sources/App/AppExtensions/Share/ShareExtensionScene.swift
@@ -1,9 +1,5 @@
-import Foundation
-import Models
-import Services
 import SwiftUI
 import Utils
-import Views
 
 public extension PlatformViewController {
   static func makeShareExtensionController(extensionContext: NSExtensionContext?) -> PlatformViewController {
@@ -16,60 +12,5 @@ public extension PlatformViewController {
       hostingController.view.layer.isOpaque = false
     #endif
     return hostingController
-  }
-}
-
-public class ShareExtensionViewModel: ObservableObject {
-  @Published var title: String?
-  @Published var debugText: String?
-
-  let saveService = ExtensionSaveService()
-
-  func handleReadNowAction(requestId: String, extensionContext: NSExtensionContext?) {
-    #if os(iOS)
-      let deepLinkUrl = NSURL(string: "omnivore://shareExtensionRequestID/\(requestId)")
-      if let application = UIApplication.value(forKeyPath: #keyPath(UIApplication.shared)) as? UIApplication {
-        application.perform(NSSelectorFromString("openURL:"), with: deepLinkUrl)
-      }
-    #else
-      let deepLinkUrl = URL(string: "omnivore://shareExtensionRequestID/\(requestId)")
-      let workspace = NSWorkspace.value(forKeyPath: #keyPath(NSWorkspace.shared)) as? NSWorkspace
-      if let workspace = workspace, let deepLinkUrl = deepLinkUrl {
-        workspace.open(deepLinkUrl)
-      }
-    #endif
-    extensionContext?.completeRequest(returningItems: [], completionHandler: nil)
-  }
-
-  func savePage(extensionContext: NSExtensionContext?, shareExtensionViewModel: ShareExtensionChildViewModel) {
-    if let extensionContext = extensionContext {
-      saveService.save(extensionContext, shareExtensionViewModel: shareExtensionViewModel)
-    } else {
-      DispatchQueue.main.async {
-        shareExtensionViewModel.status = .failed(error: .unknown(description: "Internal Error"))
-      }
-    }
-  }
-}
-
-struct ShareExtensionView: View {
-  let extensionContext: NSExtensionContext?
-  @StateObject private var viewModel = ShareExtensionViewModel()
-  @StateObject private var childViewModel = ShareExtensionChildViewModel()
-
-  var body: some View {
-    ShareExtensionChildView(
-      viewModel: childViewModel,
-      onAppearAction: {
-        viewModel.savePage(
-          extensionContext: extensionContext,
-          shareExtensionViewModel: childViewModel
-        )
-      },
-      readNowButtonAction: { viewModel.handleReadNowAction(requestId: $0, extensionContext: extensionContext) },
-      dismissButtonTappedAction: { _, _ in
-        extensionContext?.completeRequest(returningItems: [], completionHandler: nil)
-      }
-    )
   }
 }
